requirements.txt
asgiref==3.3.4
Django==3.2
pytz==2021.1
redis3==3.5.2.2
schedule==1.1.0
sqlparse==0.4.1
